Refactor tests in test/Main.hs into separate test suites.
