+++ test/Streamly/Test/Prelude/WAsync.hs
@@ -9,7 +9,11 @@
 
 module Streamly.Test.Prelude.WAsync where
 
-#if __GLASGOW_HASKELL__ < 808
+#ifdef DEVBUILD
+import Control.Concurrent ( threadDelay )
+#endif
+import Data.List (sort)
+#if !(MIN_VERSION_base(4,11,0))
 import Data.Semigroup ((<>))
 #endif
 import Test.Hspec.QuickCheck
@@ -18,7 +22,7 @@ import Test.Hspec as H
 import Streamly.Prelude
 import qualified Streamly.Prelude as S
 
-import Streamly.Test.Prelude
+import Streamly.Test.Prelude.Common
 
 main :: IO ()
 main = hspec
@@ -43,6 +47,14 @@ main = hspec
 
     describe "Monoid operations" $ do
         wAsyncOps $ monoidOps "wAsyncly" mempty sortEq
+
+    describe "WAsync loops" $ loops wAsyncly sort sort
+
+    describe "Bind and Monoidal composition combinations" $ do
+        wAsyncOps $ bindAndComposeSimpleOps "WAsync" sortEq
+        wAsyncOps $ bindAndComposeHierarchyOps "WAsync"
+        wAsyncOps $ nestTwoStreams "WAsync" sort sort
+        wAsyncOps $ nestTwoStreamsApp "WAsync" sort sort
 
     describe "Semigroup operations" $ do
         wAsyncOps $ semigroupOps "wAsyncly" sortEq
@@ -81,3 +93,70 @@ main = hspec
         wAsyncOps $ eliminationOps folded "wAsyncly folded"
         wAsyncOps $ eliminationOpsWord8 S.fromFoldable "wAsyncly"
         wAsyncOps $ eliminationOpsWord8 folded "wAsyncly folded"
+    
+    -- describe "WAsync interleaved (<>) ordering check" $
+    --     interleaveCheck wAsyncly (<>)
+    -- describe "WAsync interleaved mappend ordering check" $
+    --     interleaveCheck wAsyncly mappend
+
+    -- XXX this keeps failing intermittently, need to investigate
+    -- describe "WAsync (<>) time order check" $
+    --     parallelCheck wAsyncly (<>)
+    -- describe "WAsync mappend time order check" $
+    --     parallelCheck wAsyncly mappend
+
+    describe "Composed MonadThrow wAsyncly" $ composeWithMonadThrow wAsyncly
+
+    -- Ad-hoc tests
+    it "takes n from stream of streams" (takeCombined 3 wAsyncly)
+
+#ifdef DEVBUILD
+    let timed :: (IsStream t, Monad (t IO)) => Int -> t IO Int
+        timed x = S.yieldM (threadDelay (x * 100000)) >> return x
+
+    -- These are not run parallely because the timing gets affected
+    -- unpredictably when other tests are running on the same machine.
+    --
+    -- Also, they fail intermittently due to scheduling delays, so not run on
+    -- CI machines.
+    describe "Nested parallel and serial compositions" $ do
+        let t = timed
+            p = wAsyncly
+            s = serially
+        {-
+        -- This is not correct, the result can also be [4,4,8,0,8,0,2,2]
+        -- because of parallelism of [8,0] and [8,0].
+        it "Nest <|>, <>, <|> (1)" $
+            let t = timed
+             in toListSerial (
+                    ((t 8 <|> t 4) <> (t 2 <|> t 0))
+                <|> ((t 8 <|> t 4) <> (t 2 <|> t 0)))
+            `shouldReturn` ([4,4,8,8,0,0,2,2])
+        -}
+        it "Nest <|>, <>, <|> (2)" $
+            (S.toList . wAsyncly) (
+                   s (p (t 4 <> t 8) <> p (t 1 <> t 2))
+                <> s (p (t 4 <> t 8) <> p (t 1 <> t 2)))
+            `shouldReturn` ([4,4,8,8,1,1,2,2])
+        -- FIXME: These two keep failing intermittently on Mac OS X
+        -- Need to examine and fix the tests.
+        {-
+        it "Nest <|>, <=>, <|> (1)" $
+            let t = timed
+             in toListSerial (
+                    ((t 8 <|> t 4) <=> (t 2 <|> t 0))
+                <|> ((t 9 <|> t 4) <=> (t 2 <|> t 0)))
+            `shouldReturn` ([4,4,0,0,8,2,9,2])
+        it "Nest <|>, <=>, <|> (2)" $
+            let t = timed
+             in toListSerial (
+                    ((t 4 <|> t 8) <=> (t 1 <|> t 2))
+                <|> ((t 4 <|> t 9) <=> (t 1 <|> t 2)))
+            `shouldReturn` ([4,4,1,1,8,2,9,2])
+        -}
+        it "Nest <|>, <|>, <|>" $
+            (S.toList . wAsyncly) (
+                    ((t 4 <> t 8) <> (t 0 <> t 2))
+                <> ((t 4 <> t 8) <> (t 0 <> t 2)))
+            `shouldReturn` ([0,0,2,2,4,4,8,8])
+#endif
